requirements.txt
# CCGuide Dependencies
# Claude Code AI Guide system requirements

# Core dependencies
google-generativeai>=0.8.0
python-dotenv>=1.0.0

# Optional dependencies for enhanced functionality
requests>=2.31.0
pyyaml>=6.0

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0